// src/types/languages.ts
export type NameLanguageField = 'nameEn' | 'nameRu' | 'nameUz' | 'nameKa'
export type AuthorLanguageField =
  | 'authorEn'
  | 'authorRu'
  | 'authorUz'
  | 'authorKa'
export type AudioLanguageField =
  | 'audioFileEn'
  | 'audioFileRu'
  | 'audioFileUz'
  | 'audioFileKa'
export type DescriptionLanguageField =
  | 'descriptionEn'
  | 'descriptionRu'
  | 'descriptionUz'
  | 'descriptionKa'
export type LanguageField =
  | NameLanguageField
  | AuthorLanguageField
  | AudioLanguageField
  | DescriptionLanguageField